+++ app/src/components/Headers/Album.tsx
@@ -32,7 +32,6 @@ export default function AlbumHeader({ album }: { album: AlbumType }) {
             label={album.audioQuality.toLowerCase()}
             size="small"
             style={{
-              margin: "0.3rem 0.2rem",
               color:
                 album?.audioQuality?.toLowerCase() === "lossless"
                   ? theme.palette.common.white
@@ -43,23 +42,20 @@ export default function AlbumHeader({ album }: { album: AlbumType }) {
                   : theme.palette.primary.main,
             }}
           />
-          <Chip
-            label={`${album.numberOfTracks} tracks`}
-            size="small"
-            style={{ margin: "0.3rem 0.2rem" }}
-          />
+          <Chip label={`${album.numberOfTracks} tracks`} size="small" />
           <Chip
             label={`${Math.round(album.duration / 60)} min`}
             size="small"
-            style={{ margin: "0.3rem 0.2rem" }}
             variant="outlined"
           />
           <Chip
             label={`${new Date(album.releaseDate).getFullYear()}`}
             size="small"
             variant="outlined"
-            style={{ margin: "0.3rem 0.2rem" }}
           />
+          {album.explicit && (
+            <Chip label="Explicit" size="small" variant="outlined" />
+          )}
           <DownloadButton
             item={album}
             id={album.id}